+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Added support for zipapps.
 - Added support for `PYTHONPATH`.
+- Added support for `-I` in Python 3.x variants.
 
 ## Version 12